clean up code changes
