+++ fields.go
@@ -1,9 +1,13 @@
 package conf
 
 import (
+	"encoding"
 	"fmt"
 	"reflect"
+	"strconv"
 	"strings"
+	"time"
+	"unicode"
 )
 
 // field maintains information about a field in the configuration struct.
@@ -13,8 +17,8 @@ type field struct {
 	field   reflect.Value
 	options fieldOptions
 
-	// Important for flag parsing or any other source where booleans might be
-	// treated specially.
+	// Important for flag parsing or any other source where
+	// booleans might be treated specially.
 	boolField bool
 
 	// For usage ...  TODO: I need more.
@@ -112,8 +116,8 @@ func extractFields(prefix []string, target interface{}) ([]field, error) {
 			fields = append(fields, field{
 				name:      fieldName,
 				key:       fieldKey,
-				flagName:  getFlagName(fieldKey),
-				envName:   getEnvName(fieldKey),
+				flagName:  strings.ToLower(strings.Join(fieldKey, `-`)),
+				envName:   strings.ToLower(strings.Join(fieldKey, `-`)),
 				field:     f,
 				options:   fieldOpts,
 				boolField: f.Kind() == reflect.Bool,
@@ -171,4 +175,204 @@ func parseTag(tagStr string) (fieldOptions, error) {
 	}
 
 	return f, nil
+}
+
+// camelSplit takes a string based on camel case and splits it.
+func camelSplit(src string) []string {
+	if src == "" {
+		return []string{}
+	}
+	if len(src) < 2 {
+		return []string{src}
+	}
+
+	runes := []rune(src)
+
+	lastClass := charClass(runes[0])
+	lastIdx := 0
+	out := []string{}
+
+	// Split into fields based on class of unicode character.
+	for i, r := range runes {
+		class := charClass(r)
+
+		// If the class has transitioned.
+		if class != lastClass {
+
+			// If going from uppercase to lowercase, we want to retain the last
+			// uppercase letter for names like FOOBar, which should split to
+			// FOO Bar.
+			switch {
+			case lastClass == classUpper && class != classNumber:
+				if i-lastIdx > 1 {
+					out = append(out, string(runes[lastIdx:i-1]))
+					lastIdx = i - 1
+				}
+			default:
+				out = append(out, string(runes[lastIdx:i]))
+				lastIdx = i
+			}
+		}
+
+		if i == len(runes)-1 {
+			out = append(out, string(runes[lastIdx:]))
+		}
+		lastClass = class
+	}
+
+	return out
+}
+
+func processField(value string, field reflect.Value) error {
+	typ := field.Type()
+
+	// Look for a Set method.
+	setter := setterFrom(field)
+	if setter != nil {
+		return setter.Set(value)
+	}
+
+	if t := textUnmarshaler(field); t != nil {
+		return t.UnmarshalText([]byte(value))
+	}
+
+	if b := binaryUnmarshaler(field); b != nil {
+		return b.UnmarshalBinary([]byte(value))
+	}
+
+	if typ.Kind() == reflect.Ptr {
+		typ = typ.Elem()
+		if field.IsNil() {
+			field.Set(reflect.New(typ))
+		}
+		field = field.Elem()
+	}
+
+	switch typ.Kind() {
+	case reflect.String:
+		field.SetString(value)
+	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
+		var (
+			val int64
+			err error
+		)
+		if field.Kind() == reflect.Int64 && typ.PkgPath() == "time" && typ.Name() == "Duration" {
+			var d time.Duration
+			d, err = time.ParseDuration(value)
+			val = int64(d)
+		} else {
+			val, err = strconv.ParseInt(value, 0, typ.Bits())
+		}
+		if err != nil {
+			return err
+		}
+
+		field.SetInt(val)
+	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
+		val, err := strconv.ParseUint(value, 0, typ.Bits())
+		if err != nil {
+			return err
+		}
+		field.SetUint(val)
+	case reflect.Bool:
+		val, err := strconv.ParseBool(value)
+		if err != nil {
+			return err
+		}
+		field.SetBool(val)
+	case reflect.Float32, reflect.Float64:
+		val, err := strconv.ParseFloat(value, typ.Bits())
+		if err != nil {
+			return err
+		}
+		field.SetFloat(val)
+	case reflect.Slice:
+		vals := strings.Split(value, ",")
+		sl := reflect.MakeSlice(typ, len(vals), len(vals))
+		for i, val := range vals {
+			err := processField(val, sl.Index(i))
+			if err != nil {
+				return err
+			}
+		}
+		field.Set(sl)
+	case reflect.Map:
+		mp := reflect.MakeMap(typ)
+		if len(strings.TrimSpace(value)) != 0 {
+			pairs := strings.Split(value, ",")
+			for _, pair := range pairs {
+				kvpair := strings.Split(pair, ":")
+				if len(kvpair) != 2 {
+					return fmt.Errorf("invalid map item: %q", pair)
+				}
+				k := reflect.New(typ.Key()).Elem()
+				err := processField(kvpair[0], k)
+				if err != nil {
+					return err
+				}
+				v := reflect.New(typ.Elem()).Elem()
+				err = processField(kvpair[1], v)
+				if err != nil {
+					return err
+				}
+				mp.SetMapIndex(k, v)
+			}
+		}
+		field.Set(mp)
+	}
+	return nil
+}
+
+func interfaceFrom(field reflect.Value, fn func(interface{}, *bool)) {
+
+	// It may be impossible for a struct field to fail this check.
+	if !field.CanInterface() {
+		return
+	}
+
+	var ok bool
+	fn(field.Interface(), &ok)
+	if !ok && field.CanAddr() {
+		fn(field.Addr().Interface(), &ok)
+	}
+}
+
+// Setter is implemented by types can self-deserialize values.
+// Any type that implements flag.Value also implements Setter.
+type Setter interface {
+	Set(value string) error
+}
+
+func setterFrom(field reflect.Value) (s Setter) {
+	interfaceFrom(field, func(v interface{}, ok *bool) { s, *ok = v.(Setter) })
+	return s
+}
+
+func textUnmarshaler(field reflect.Value) (t encoding.TextUnmarshaler) {
+	interfaceFrom(field, func(v interface{}, ok *bool) { t, *ok = v.(encoding.TextUnmarshaler) })
+	return t
+}
+
+func binaryUnmarshaler(field reflect.Value) (b encoding.BinaryUnmarshaler) {
+	interfaceFrom(field, func(v interface{}, ok *bool) { b, *ok = v.(encoding.BinaryUnmarshaler) })
+	return b
+}
+
+const (
+	classLower int = iota
+	classUpper
+	classNumber
+	classOther
+)
+
+func charClass(r rune) int {
+	switch {
+	case unicode.IsLower(r):
+		return classLower
+	case unicode.IsUpper(r):
+		return classUpper
+	case unicode.IsDigit(r):
+		return classNumber
+	}
+	return classOther
 }
