--- process.go
@@ -1,145 +1,0 @@
-package conf
-
-import (
-	"encoding"
-	"fmt"
-	"reflect"
-	"strconv"
-	"strings"
-	"time"
-)
-
-func processField(value string, field reflect.Value) error {
-	typ := field.Type()
-
-	// Look for a Set method.
-	setter := setterFrom(field)
-	if setter != nil {
-		return setter.Set(value)
-	}
-
-	if t := textUnmarshaler(field); t != nil {
-		return t.UnmarshalText([]byte(value))
-	}
-
-	if b := binaryUnmarshaler(field); b != nil {
-		return b.UnmarshalBinary([]byte(value))
-	}
-
-	if typ.Kind() == reflect.Ptr {
-		typ = typ.Elem()
-		if field.IsNil() {
-			field.Set(reflect.New(typ))
-		}
-		field = field.Elem()
-	}
-
-	switch typ.Kind() {
-	case reflect.String:
-		field.SetString(value)
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		var (
-			val int64
-			err error
-		)
-		if field.Kind() == reflect.Int64 && typ.PkgPath() == "time" && typ.Name() == "Duration" {
-			var d time.Duration
-			d, err = time.ParseDuration(value)
-			val = int64(d)
-		} else {
-			val, err = strconv.ParseInt(value, 0, typ.Bits())
-		}
-		if err != nil {
-			return err
-		}
-
-		field.SetInt(val)
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-		val, err := strconv.ParseUint(value, 0, typ.Bits())
-		if err != nil {
-			return err
-		}
-		field.SetUint(val)
-	case reflect.Bool:
-		val, err := strconv.ParseBool(value)
-		if err != nil {
-			return err
-		}
-		field.SetBool(val)
-	case reflect.Float32, reflect.Float64:
-		val, err := strconv.ParseFloat(value, typ.Bits())
-		if err != nil {
-			return err
-		}
-		field.SetFloat(val)
-	case reflect.Slice:
-		vals := strings.Split(value, ",")
-		sl := reflect.MakeSlice(typ, len(vals), len(vals))
-		for i, val := range vals {
-			err := processField(val, sl.Index(i))
-			if err != nil {
-				return err
-			}
-		}
-		field.Set(sl)
-	case reflect.Map:
-		mp := reflect.MakeMap(typ)
-		if len(strings.TrimSpace(value)) != 0 {
-			pairs := strings.Split(value, ",")
-			for _, pair := range pairs {
-				kvpair := strings.Split(pair, ":")
-				if len(kvpair) != 2 {
-					return fmt.Errorf("invalid map item: %q", pair)
-				}
-				k := reflect.New(typ.Key()).Elem()
-				err := processField(kvpair[0], k)
-				if err != nil {
-					return err
-				}
-				v := reflect.New(typ.Elem()).Elem()
-				err = processField(kvpair[1], v)
-				if err != nil {
-					return err
-				}
-				mp.SetMapIndex(k, v)
-			}
-		}
-		field.Set(mp)
-	}
-	return nil
-}
-
-func interfaceFrom(field reflect.Value, fn func(interface{}, *bool)) {
-
-	// It may be impossible for a struct field to fail this check.
-	if !field.CanInterface() {
-		return
-	}
-
-	var ok bool
-	fn(field.Interface(), &ok)
-	if !ok && field.CanAddr() {
-		fn(field.Addr().Interface(), &ok)
-	}
-}
-
-// Setter is implemented by types can self-deserialize values.
-// Any type that implements flag.Value also implements Setter.
-type Setter interface {
-	Set(value string) error
-}
-
-func setterFrom(field reflect.Value) (s Setter) {
-	interfaceFrom(field, func(v interface{}, ok *bool) { s, *ok = v.(Setter) })
-	return s
-}
-
-func textUnmarshaler(field reflect.Value) (t encoding.TextUnmarshaler) {
-	interfaceFrom(field, func(v interface{}, ok *bool) { t, *ok = v.(encoding.TextUnmarshaler) })
-	return t
-}
-
-func binaryUnmarshaler(field reflect.Value) (b encoding.BinaryUnmarshaler) {
-	interfaceFrom(field, func(v interface{}, ok *bool) { b, *ok = v.(encoding.BinaryUnmarshaler) })
-	return b
-}
